+++ tests/e2e/test_online_features.py
@@ -23,6 +23,7 @@ from feast import (
 from feast.data_format import AvroFormat, ParquetFormat
 from feast.pyspark.abc import SparkJobStatus
 from feast.wait import wait_retry_backoff
+from tests.e2e.utils.common import create_schema, start_job, stop_job
 from tests.e2e.utils.kafka import check_consumer_exist, ingest_and_retrieve
 
 
@@ -45,27 +46,6 @@ def test_offline_ingestion(
 
     feature_table = FeatureTable(
         name="drivers",
-        entities=["s2id"],
-        features=[Feature("unique_drivers", ValueType.INT64)],
-        batch_source=batch_source,
-    )
-
-    feast_client.apply(entity)
-    feast_client.apply(feature_table)
-
-    original = generate_data()
-    feast_client.ingest(feature_table, original)  # write to batch (offline) storage
-
-    ingest_and_verify(feast_client, feature_table, original)
-
-
-def test_offline_ingestion_long_table_name(
-    feast_client: Client, batch_source: Union[BigQuerySource, FileSource]
-):
-    entity = Entity(name="s2id", description="S2id", value_type=ValueType.INT64,)
-
-    feature_table = FeatureTable(
-        name="just1a2featuretable3with4a5really6really7really8really9really10really11really12long13name",
         entities=["s2id"],
         features=[Feature("unique_drivers", ValueType.INT64)],
         batch_source=batch_source,
@@ -193,13 +173,6 @@ def ingest_and_verify(
         original.event_timestamp.max().to_pydatetime() + timedelta(seconds=1),
     )
     assert job.get_feature_table()[:63] == feature_table.name[:63]
-    all_job_ids = [
-        job.get_id()
-        for job in feast_client.list_jobs(
-            include_terminated=True, table_name=feature_table.name
-        )
-    ]
-    assert job.get_id() in all_job_ids
 
     wait_retry_backoff(
         lambda: (None, job.get_status() == SparkJobStatus.COMPLETED), 180
@@ -217,6 +190,31 @@ def ingest_and_verify(
             columns={"unique_drivers": f"{feature_table.name}:unique_drivers"}
         ),
     )
+
+
+def test_list_jobs_long_table_name(feast_client: Client, kafka_server, pytestconfig):
+    kafka_broker = f"{kafka_server[0]}:{kafka_server[1]}"
+    topic_name = f"avro-{uuid.uuid4()}"
+
+    entity, feature_table = create_schema(
+        kafka_broker,
+        topic_name,
+        "just1a2featuretable3with4a5really6really7really8really9really10really11really12long13name",
+        avro_schema(),
+    )
+
+    feast_client.apply(entity)
+    feast_client.apply(feature_table)
+
+    job = start_job(feast_client, feature_table, pytestconfig)
+    all_job_ids = [
+        job.get_id()
+        for job in feast_client.list_jobs(
+            include_terminated=True, table_name=feature_table.name
+        )
+    ]
+    assert job.get_id() in all_job_ids
+    stop_job(job, feast_client, feature_table)
 
 
 def avro_schema():
