+++ tests/e2e/test_validation.py
@@ -7,20 +7,11 @@ import pandas as pd
 import pytest
 from great_expectations.dataset import PandasDataset
 
-from feast import (
-    Client,
-    Entity,
-    Feature,
-    FeatureTable,
-    FileSource,
-    KafkaSource,
-    ValueType,
-)
+from feast import Client
 from feast.contrib.validation.ge import apply_validation, create_validation_udf
-from feast.data_format import AvroFormat, ParquetFormat
-from feast.pyspark.abc import SparkJobStatus
 from feast.wait import wait_retry_backoff
 from tests.e2e.fixtures.statsd_stub import StatsDServer
+from tests.e2e.utils.common import create_schema, start_job, stop_job
 from tests.e2e.utils.kafka import check_consumer_exist, ingest_and_retrieve
 
 
@@ -44,50 +35,13 @@ def generate_test_data():
     return df
 
 
-def create_schema(kafka_broker, topic_name, feature_table_name):
-    entity = Entity(name="key", description="Key", value_type=ValueType.INT64)
-    feature_table = FeatureTable(
-        name=feature_table_name,
-        entities=["key"],
-        features=[Feature("num", ValueType.INT64), Feature("set", ValueType.STRING)],
-        batch_source=FileSource(
-            event_timestamp_column="event_timestamp",
-            file_format=ParquetFormat(),
-            file_url="/dev/null",
-        ),
-        stream_source=KafkaSource(
-            event_timestamp_column="event_timestamp",
-            bootstrap_servers=kafka_broker,
-            message_format=AvroFormat(avro_schema()),
-            topic=topic_name,
-        ),
-    )
-    return entity, feature_table
-
-
-def start_job(feast_client: Client, feature_table: FeatureTable, pytestconfig):
-    if pytestconfig.getoption("scheduled_streaming_job"):
-        return
-
-    job = feast_client.start_stream_to_online_ingestion(feature_table)
-    wait_retry_backoff(
-        lambda: (None, job.get_status() == SparkJobStatus.IN_PROGRESS), 120
-    )
-    return job
-
-
-def stop_job(job, feast_client: Client, feature_table: FeatureTable):
-    if job:
-        job.cancel()
-    else:
-        feast_client.delete_feature_table(feature_table.name)
-
-
 def test_validation_with_ge(feast_client: Client, kafka_server, pytestconfig):
     kafka_broker = f"{kafka_server[0]}:{kafka_server[1]}"
     topic_name = f"avro-{uuid.uuid4()}"
 
-    entity, feature_table = create_schema(kafka_broker, topic_name, "validation_ge")
+    entity, feature_table = create_schema(
+        kafka_broker, topic_name, "validation_ge", avro_schema()
+    )
     feast_client.apply_entity(entity)
     feast_client.apply_feature_table(feature_table)
 
@@ -153,7 +107,7 @@ def test_validation_reports_metrics(
     topic_name = f"avro-{uuid.uuid4()}"
 
     entity, feature_table = create_schema(
-        kafka_broker, topic_name, "validation_ge_metrics"
+        kafka_broker, topic_name, "validation_ge_metrics", avro_schema()
     )
     feast_client.apply_entity(entity)
     feast_client.apply_feature_table(feature_table)
